refactor(cli): extract debug server into a feature module
Move the debug HTTP server out of `utils/` (it's a subsystem, not a
small utility) into `cli/debug-server/` and decompose the former
167-line `createDebugServer` god-factory:

- `create-debug-server.ts` — lifecycle only (lock reuse, listen, lock
  write, exit cleanup).
- `ingest-request-listener.ts` — CORS + a route table over the health
  check and `/ingest/:sessionId` POST/GET/DELETE, with a `writeJson`
  helper replacing the repeated writeHead/end(JSON.stringify) blocks.
- `debug-session-store.ts` — per-session state, id parsing/validation,
  and bounded dedup memory.
- `server-lock.ts` / `ping-server.ts` — moved unchanged.

No behavior change; addresses the thermo code-quality P1.
