+++ packages/react-doctor/src/cli/debug-server/index.ts
@@ -1,0 +1,6 @@
+export { createDebugServer } from "./create-debug-server.js";
+export type {
+  DebugServerInfo,
+  DebugServerOptions,
+  DebugServerResult,
+} from "./create-debug-server.js";
